+++ Client/menus/menuBack/menuBack.hpp
@@ -1,12 +1,11 @@
 #pragma once
-
 #include "graphics.hpp"
 
 class BackgroundRect {
 public:
     virtual void renderBack() = 0;
-    virtual void setBackWidth(unsigned short width) = 0;
-    virtual unsigned short getBackWidth() = 0;
+    virtual void setBackWidth(int width) = 0;
+    virtual int getBackWidth() = 0;
 };
 
 class MenuBack : public BackgroundRect {
@@ -15,6 +14,6 @@ class MenuBack : public BackgroundRect {
 public:
     void init();
     void renderBack() override;
-    void setBackWidth(unsigned short width) override { back_rect.setWidth(width); }
-    unsigned short getBackWidth() override { return back_rect.getWidth(); }
+    void setBackWidth(int width) override { back_rect.setWidth(width); }
+    int getBackWidth() override { return back_rect.getWidth(); }
 };
