+++ Libraries/graphics/centeredObject.cpp
@@ -1,12 +1,12 @@
 #include "graphics-internal.hpp"
 
-gfx::_CenteredObject::_CenteredObject(short x, short y, Orientation orientation) : orientation(orientation), x(x), y(y) {}
+gfx::_CenteredObject::_CenteredObject(int x, int y, Orientation orientation) : orientation(orientation), x(x), y(y) {}
 
-short gfx::_CenteredObject::getTranslatedX() const {
+int gfx::_CenteredObject::getTranslatedX() const {
     return x + getWindowWidth() * orientation.x - getWidth() * orientation.x;
 }
 
-short gfx::_CenteredObject::getTranslatedY() const {
+int gfx::_CenteredObject::getTranslatedY() const {
     return y + getWindowHeight() * orientation.y - getHeight() * orientation.y;
 }
